:recycle: refactor: 컴포넌트 폴더명 UpperCamelCase로 변경
