+++ frontend/src/components/GroupArticleCard/GroupArticleCard.stories.tsx
