+++ frontend/src/components/GroupArticleCard/styles.ts
